Move remaining git2 dependencies into git.rs.
This finally confines the dependency on git2 in the single source,
making the rest of the logic abstract (working with types from
commit.rs).

However, there are few problems in the resulting code:

* git.rs currently exposes only concrete types, so introducing support
for other VCS types is still problematic. It is possible to use few
levels of traits with dynamic dispatch, but I don't feel that this is a
good solution, so don't use it unless really required;
* the MessageInfo is forced to own the subject message to make the
git.rs interface convenient. It would be nice to discover the
alternative approaches.