Update TODOs to the actual state.
Remove TODOs and XXXes about owned Strings in structures - the major
load of CPU and memory consumption come from diffs calculation, so
making a stress on two additional allocations/copies per commit is quite
misleading.

Aldo, add a TODO for commit.rs tests as a reminder that this module is
not yet fully covered.